+++ BetterBulldozer/UI/src/mods/betterBulldozerSections/betterBulldozerSections.tsx
@@ -4,22 +4,39 @@ import { bindValue, trigger, useValue } from "cs2/api";
 import { VanillaComponentResolver } from "../VanillaComponentResolver/VanillaComponentResolver";
 import mod from "../../../mod.json";
 import spacesSrc from "./AreaToolBetterBulldozer.svg";
-import bypassConfirmationSrc from "./BypassQuestionmarkBB.svg";
-import gameplayManipulationSrc from "./CubeSimulationBB.svg";
 import staticObjectMarkersSrc from "./MarkersIconBetterBulldozer.svg";
-import lanesSrc from "./NetworkBB.svg";
-import networkMarkersSrc from "./DottedLinesMarkersBB.svg";
 import surfacesSrc from "./SurfaceIconBetterBulldozer.svg";
 import { tool } from "cs2/bindings";
-import { FOCUS_DISABLED, FocusSymbol, PanelSectionRow } from "cs2/ui";
 
 // These establishes the binding with C# side. Without C# side game ui will crash.
-// export const bulldozeToolActive$ = bindValue<boolean>(mod.id, 'BulldozeToolActive');
 export const gameplayManipulation$ = bindValue<boolean>(mod.id, 'GameplayManipulation');
 export const bypassConfirmation$ = bindValue<boolean>(mod.id, 'BypassConfirmation');
 export const raycastTarget$ = bindValue<number>(mod.id, 'RaycastTarget');
 export const areasFilter$ = bindValue<number>(mod.id, 'AreasFilter');
 export const markersFilter$ = bindValue<number>(mod.id, 'MarkersFilter');
+
+// These contain the coui paths to Unified Icon Library svg assets
+export const couiStandard =                         "coui://uil/Standard/";
+export const gameplayManipulationSrc =         couiStandard +  "CubeSimulation.svg";
+export const bypassConfirmationSrc =           couiStandard +  "BypassQuestionmark.svg";
+export const lanesSrc =                         couiStandard + "Network.svg";
+export const networkMarkersSrc =                couiStandard + "DottedLinesMarkers.svg";
+export const iconsSrc =                         couiStandard + "ArrowDownTriangle.svg";
+
+
+// Saving strings for events and translations.
+export const surfacesID =              "SurfacesFilterButton";
+export const spacesID =                "SpacesFilterButton";
+export const staticObjectsID =         "StaticObjectsFilterButton";
+export const networksFilterID =        "NetworksFilterButton";
+export const gameplayManipulationID =  "GameplayManipulationButton";
+export const bypassConfirmationID =    "BypassConfirmationButton";
+export const raycastMarkersID =        "RaycastMarkersButton";
+export const raycastAreasID =          "RaycastAreasButton";
+export const lanesID =                 "RaycastLanesButton";
+export const iconsID =                 "RaycastIconsButton";
+export const tooltipDescriptionPrefix ="YY_BETTER_BULLDOZER_DESCRIPTION.";
+export const sectionTitlePrefix =      "YY_BETTER_BULLDOZER.";
 
 // This functions trigger an event on C# side and C# designates the method to implement.
 export function handleClick(eventName: string) {
@@ -40,19 +57,6 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
         const areasFilter = useValue(areasFilter$);
         const markersFilter = useValue(markersFilter$);
         
-        // Saving strings for events and translations.
-        const surfacesID =              "SurfacesFilterButton";
-        const spacesID =                "SpacesFilterButton";
-        const staticObjectsID =         "StaticObjectsFilterButton";
-        const networksFilterID =        "NetworksFilterButton";
-        const gameplayManipulationID =  "GameplayManipulationButton";
-        const bypassConfirmationID =    "BypassConfirmationButton";
-        const raycastMarkersID =        "RaycastMarkersButton";
-        const raycastAreasID =          "RaycastAreasButton";
-        const lanesID =                 "RaycastLanesButton";
-        const tooltipDescriptionPrefix ="YY_BETTER_BULLDOZER_DESCRIPTION.";
-        const sectionTitlePrefix =      "YY_BETTER_BULLDOZER.";
-
         // translation handling. Translates using locale keys that are defined in C# or fallback string here.
         const { translate } = useLocalization();
         const filterSectionTitle =          translate(sectionTitlePrefix + "Filter",                        "Filter");
@@ -65,6 +69,7 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
         const raycastMarkersTooltip =       translate(tooltipDescriptionPrefix + raycastMarkersID,          "Shows and EXCLUSIVELY targets static object markers or invisible networks. With this enabled you can demolish invisible networks, invisible parking decals, various spots, points, and spawners, but SAVE FIRST! You cannot demolish these within buildings.");
         const raycastAreasTooltip =         translate(tooltipDescriptionPrefix + raycastAreasID,            "Makes the bulldozer EXCLUSIVELY target surfaces or spaces inside or outside of buildings so you can remove them in one click. You must turn this off to bulldoze anything else.");
         const lanesTooltip =                translate(tooltipDescriptionPrefix + lanesID,                   "For removing standalone lanes such as interconnected fences, interconnected hedges, linear street markings, vehicle and pedestrian lanes. Trying to target those inside networks will remove the network! You cannot create these in-game without a mod for it.");
+        const iconsTooltip =                translate(tooltipDescriptionPrefix + iconsID,                   "For removing floating icons.");
         const toolModeTitle =               translate("Toolbar.TOOL_MODE_TITLE", "Tool Mode");
 
         // These convert integer casts of Enums into booleans.
@@ -75,6 +80,7 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
         const spacesFilter : boolean = areasFilter == 8;
         const staticObjectMarkersFilter : boolean = markersFilter == 2; 
         const markerNetworkFilter : boolean = markersFilter == 8;
+        const raycastingIcons : boolean = raycastTarget == 4;
 
         // This gets the original component that we may alter and return.
         var result : JSX.Element = Component();
@@ -105,9 +111,10 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
                         : <></>
                     }
                     <VanillaComponentResolver.instance.Section title={toolModeTitle}>
-                            <VanillaComponentResolver.instance.ToolButton  selected={raycastingLanes}       tooltip={lanesTooltip}                  onSelect={() => handleClick(lanesID)}                   src={lanesSrc}                 focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
                             <VanillaComponentResolver.instance.ToolButton  selected={gameplayManipulation}  tooltip={gameplayManipulationTooltip}   onSelect={() => handleClick(gameplayManipulationID)}    src={gameplayManipulationSrc}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
                             <VanillaComponentResolver.instance.ToolButton  selected={bypassConfirmation}    tooltip={bypassConfirmationTooltip}     onSelect={() => handleClick(bypassConfirmationID)}      src={bypassConfirmationSrc}    focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={raycastingIcons}       tooltip={iconsTooltip}                  onSelect={() => handleClick(iconsID)}                   src={iconsSrc}                 focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={raycastingLanes}       tooltip={lanesTooltip}                  onSelect={() => handleClick(lanesID)}                   src={lanesSrc}                 focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
                             <VanillaComponentResolver.instance.ToolButton  selected={raycastingMarkers}     tooltip={raycastMarkersTooltip}         onSelect={() => handleClick(raycastMarkersID)}          src={networkMarkersSrc}        focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
                             <VanillaComponentResolver.instance.ToolButton  selected={raycastingAreas}       tooltip={raycastAreasTooltip}           onSelect={() => handleClick(raycastAreasID)}            src={surfacesSrc}              focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
                     </VanillaComponentResolver.instance.Section>