+++ BetterBulldozer/ChangeLog.md
@@ -1,3 +1,4 @@
-﻿# Patch V1.1.1
-* Fixed errors related to Focus nodes, Thanks to Klyte.
-* Modified some icons by Chameleon TBN.
+﻿# Patch V1.1.2
+* Added Dependency for Unified Icon Library
+* Fixed Raycasting Lanes
+* Added Raycasting Icons